@@ -305,7 +305,7 @@ func TestInterpreterConvertTrace(t *testing.T) {
 			// For this test do not include interpreters.
 			noIinterpreters, _ := tracertypes.Parse("")
 
-			ctx, cancel := context.WithCancel(context.Background())
+			ctx, cancel := context.WithCancel(t.Context())
 			defer cancel()
 
 			// To test ConvertTrace we do not require all parts of processmanager.
@@ -387,7 +387,7 @@ func TestNewMapping(t *testing.T) {
 			// For this test do not include interpreters.
 			noInterpreters, _ := tracertypes.Parse("")
 
-			ctx, cancel := context.WithCancel(context.Background())
+			ctx, cancel := context.WithCancel(t.Context())
 			defer cancel()
 
 			manager, err := New(ctx,
@@ -574,7 +574,7 @@ func TestProcExit(t *testing.T) {
 			// For this test do not include interpreters.
 			noInterpreters, _ := tracertypes.Parse("")
 
-			ctx, cancel := context.WithCancel(context.Background())
+			ctx, cancel := context.WithCancel(t.Context())
 
 			manager, err := New(ctx,
 				noInterpreters,
